move compose annotations to main package
