+++ sample/simple/feature/dialog/implementation/src/main/kotlin/com/freeletics/khonshu/sample/feature/dialog/DialogScreen.kt
@@ -11,7 +11,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
-import com.freeletics.khonshu.codegen.compose.NavDestination
+import com.freeletics.khonshu.codegen.NavDestination
 import com.freeletics.khonshu.sample.feature.dialog.nav.DialogRoute
 
 @NavDestination(
