+++ sample/simple/feature/screen/implementation/src/main/kotlin/com/freeletics/khonshu/sample/feature/screen/ScreenScreen.kt
@@ -11,7 +11,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.freeletics.khonshu.codegen.compose.NavDestination
+import com.freeletics.khonshu.codegen.NavDestination
 import com.freeletics.khonshu.sample.feature.screen.nav.ScreenRoute
 
 @NavDestination(
